fixes ObjectStoreMemoryBytes field and processing
should be int64 so we can represent more than 2.2Gi
fixes bug with cmd args passing value instead of pointer addr

+++ api/v1alpha1/raycluster_types.go
@@ -55,7 +55,7 @@ type RayClusterSpec struct {
 	// ObjectStoreMemoryBytes is initial amount of memory with which to start the object store.
 	// +kubebuilder:validation:Optional
 	// +kubebuilder:validation:Minimum=78643200
-	ObjectStoreMemoryBytes *int32 `json:"objectStoreMemoryBytes"`
+	ObjectStoreMemoryBytes *int64 `json:"objectStoreMemoryBytes"`
 
 	// DashboardPort is the port used by the dashboard server.
 	// +kubebuilder:default=8265
